refactor: used external sendKeysPlugin and a11ySnapshotPlugin
