+++ package.json
@@ -69,8 +69,9 @@
         "storybook:stories:watch": "tsc --build .storybook/tsconfig.json -w",
         "test": "yarn test:focus unit",
         "test:bench": "node test/benchmark/cli.cjs",
-        "test:build": "tsc --build test/tsconfig-plugins.json && tsc --build test/tsconfig-test.json",
-        "test:ci": "yarn test:build && yarn test:start",
+        "test:build": "tsc --build test/tsconfig-test.json",
+        "test:build:plugins": "tsc --build test/tsconfig-plugins.json",
+        "test:ci": "yarn test:build && yarn test:build:plugins && yarn test:start",
         "test:focus": "yarn build && yarn test:ci --coverage --group",
         "test:start": "web-test-runner",
         "test:visual": "yarn test:visual:ci",
@@ -98,7 +99,7 @@
         "@open-wc/dev-server-hmr": "^0.1.1",
         "@open-wc/lit-helpers": "^0.3.12",
         "@open-wc/polyfills-loader": "^0.3.3",
-        "@open-wc/testing": "^2.5.32",
+        "@open-wc/testing": "^2.5.33",
         "@rollup/plugin-json": "^4.1.0",
         "@rollup/plugin-node-resolve": "^11.2.0",
         "@spectrum-css/table": "^3.0.3",
@@ -115,15 +116,15 @@
         "@typescript-eslint/eslint-plugin": "^4.16.1",
         "@typescript-eslint/parser": "^4.16.1",
         "@vaadin/router": "^1.2.0",
-        "@web/dev-server": "^0.1.11",
+        "@web/dev-server": "^0.1.16",
         "@web/dev-server-rollup": "^0.3.3",
         "@web/dev-server-storybook": "^0.3.5",
         "@web/rollup-plugin-copy": "^0.2.0",
-        "@web/rollup-plugin-html": "^1.5.0",
-        "@web/test-runner": "^0.12.19",
-        "@web/test-runner-commands": "^0.4.2",
-        "@web/test-runner-junit-reporter": "^0.4.1",
-        "@web/test-runner-playwright": "^0.8.4",
+        "@web/rollup-plugin-html": "^1.6.0",
+        "@web/test-runner": "^0.13.2",
+        "@web/test-runner-commands": "^0.4.5",
+        "@web/test-runner-junit-reporter": "^0.4.2",
+        "@web/test-runner-playwright": "^0.8.5",
         "@web/test-runner-visual-regression": "^0.6.0",
         "@webcomponents/webcomponentsjs": "^2.4.4",
         "alex": "^9.0.1",
