+++ test/plugins/send-mouse-plugin.ts
