+++ web-test-runner.config.js
@@ -10,9 +10,11 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 import { playwrightLauncher } from '@web/test-runner-playwright';
-import { a11ySnapshotPlugin } from './test/a11y-snapshot-plugin.js';
-import { sendKeysPlugin } from './test/send-keys-plugin.js';
-import { sendMousePlugin } from './test/send-mouse-plugin.js';
+import {
+    sendKeysPlugin,
+    a11ySnapshotPlugin,
+} from '@web/test-runner-commands/plugins';
+import { sendMousePlugin } from './test/plugins/send-mouse-plugin.js';
 import {
     packages,
     vrtGroups,
